fflush() saved us

+++ mainMemory.h
@@ -1,1 +1,1 @@
-short fetchMainMemory(long la, long pdpa, struct Hardware *hardware, struct Process *process);
+long fetchMainMemory(long la, long pdpa, struct Hardware *hardware, struct Process *process);
